# lib/api-zod/src/generated/types/analyzeProfileRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeProfileRequest {
  /** Raw text from uploaded resume */
  resumeText?: string;
  /** LinkedIn profile information as text */
  linkedinData?: string;
  /** Target profession or field of study (e.g. "Computer Science", "Business Administration") */
  profession: string;
  /** Preferred country for study (optional) */
  country?: string;
}