# lib/api-zod/src/generated/types/analyzeProfileResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EligibleProgram } from "./eligibleProgram";
import type { NearMatchProgram } from "./nearMatchProgram";
import type { UserProfile } from "./userProfile";

export interface AnalyzeProfileResponse {
  userProfile: UserProfile;
  /** Programs the user currently qualifies for */
  eligiblePrograms: EligibleProgram[];
  /** Programs the user is close to qualifying for */
  nearMatchPrograms: NearMatchProgram[];
  /** Brief summary of the analysis */
  summary: string;
}