# lib/api-zod/src/generated/types/eligibleProgram.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EligibleProgram {
  /** University name */
  university: string;
  /** Country where university is located */
  country: string;
  /** Program name (e.g. "MSc Computer Science") */
  program: string;
  /** Degree type (e.g. "Master's", "PhD", "Bachelor's") */
  degree: string;
  /** Program duration (e.g. "2 years") */
  duration?: string;
  /** Estimated tuition range */
  tuitionRange?: string;
  /** University or program ranking */
  ranking?: string;
  /** Key admission requirements */
  admissionRequirements: string[];
  /** Explanation of why the user is eligible */
  whyEligible: string;
  /** Typical application deadline */
  applicationDeadline?: string;
  /** Link to program info (may be approximate) */
  programUrl?: string;
}