# lib/api-zod/src/generated/types/gap.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GapType } from "./gapType";

export interface Gap {
  /** Type of gap */
  type: GapType;
  /** What is missing or insufficient */
  description: string;
  /** Actionable steps to close this gap */
  howToFix: string;
  /** Estimated time to close this gap (e.g. "3-6 months") */
  timeToFix?: string;
}