# lib/api-zod/src/generated/types/nearMatchProgram.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Gap } from "./gap";

export interface NearMatchProgram {
  /** University name */
  university: string;
  /** Country where university is located */
  country: string;
  /** Program name */
  program: string;
  /** Degree type */
  degree: string;
  /** Program duration */
  duration?: string;
  /** Estimated tuition range */
  tuitionRange?: string;
  /** University or program ranking */
  ranking?: string;
  /** List of gaps the user needs to close */
  gaps: Gap[];
  /** How close the user is (0-100 percentage) */
  matchScore: number;
  /** Typical application deadline */
  applicationDeadline?: string;
  /** Link to program info */
  programUrl?: string;
}