# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  /** GPA extracted from resume (if found) */
  extractedGpa?: string;
  /** Highest degree extracted from resume */
  extractedDegree?: string;
  /** Field of study from resume */
  extractedField?: string;
  /** Key skills extracted */
  extractedSkills: string[];
  /** Years of experience */
  extractedExperience?: string;
  /** Relevant courses mentioned */
  extractedCourses: string[];
}